requirements.txt
cvxopt==1.3.2
numpy==1.26.0
roboticstoolbox_python==1.1.0
scipy==1.11.3
spatialmath_python==1.1.8
urdfpy==0.0.22
